Crud familia admin product

+++ backend/src/models/families.ts
@@ -46,6 +46,8 @@ const importReportList: ImportReport[] = [];
 
 const csvWriterList: { [key: string]: ReturnType<typeof createObjectCsvWriter> } = {};
 
+const type = process.env.CONSUMPTION_TYPE as 'ticket' | 'product';
+
 /**
  * Get or create CSV Writer for the the city
  * @param cityId logged user unique ID
@@ -477,41 +479,43 @@ export const createFamilyWithDependents = async (
     throw { status: 412, message: 'Familia já cadastrada.' };
   }
   //verify if dependent exist
-  if (values.dependents) {
-    const verifyResponsible = values.dependents?.filter((f) => f.isResponsible);
-    if (verifyResponsible.length > 1)
-      throw { status: 412, message: 'Somente um membro responsável por familia é permitido.' };
-    if (verifyResponsible.length === 0)
-      throw { status: 412, message: 'É necessário um membro responsável por familia.' };
+  if (type === 'product') {
+    if (values.dependents) {
+      const verifyResponsible = values.dependents?.filter((f) => f.isResponsible);
+      if (verifyResponsible.length > 1)
+        throw { status: 412, message: 'Somente um membro responsável por familia é permitido.' };
+      if (verifyResponsible.length === 0)
+        throw { status: 412, message: 'É necessário um membro responsável por familia.' };
 
-    // const dependentNis = values.dependents?.map((dep) => {
-    //   return dep.nis as string;
-    // });
-    const dependentRg = values.dependents?.map((dep) => {
-      return dep.rg as string;
-    });
-    const dependentCpf = values.dependents?.map((dep) => {
-      return dep.cpf as string;
-    });
-    const dependents = await db.dependents.findAll({
-      where: Sequelize.or(
-        // { nis: { [Sequelize.Op.in]: dependentNis } },
-        { rg: { [Sequelize.Op.in]: dependentRg } },
-        { cpf: { [Sequelize.Op.in]: dependentCpf } }
-      )
-    });
-    if (dependents.length > 0) {
-      const listOfNames = dependents.map((dep) => {
-        return dep.name;
+      const dependentRg = values.dependents?.map((dep) => {
+        return dep.rg as string;
       });
-      throw {
-        status: 412,
-        message: `Dependente${listOfNames.length > 1 ? 's' : ''} ${listOfNames.toString()} já cadastrado${
-          listOfNames.length > 1 ? 's' : ''
-        }.`
-      };
+      const dependentCpf = values.dependents?.map((dep) => {
+        return dep.cpf as string;
+      });
+      const dependents = await db.dependents.findAll({
+        where: Sequelize.or(
+          // { nis: { [Sequelize.Op.in]: dependentNis } },
+          { rg: { [Sequelize.Op.in]: dependentRg } },
+          { cpf: { [Sequelize.Op.in]: dependentCpf } }
+        )
+      });
+      if (dependents.length > 0) {
+        const listOfNames = dependents.map((dep) => {
+          return dep.name;
+        });
+        throw {
+          status: 412,
+          message: `Dependente${listOfNames.length > 1 ? 's' : ''} ${listOfNames.toString()} já cadastrado${
+            listOfNames.length > 1 ? 's' : ''
+          }.`
+        };
+      }
+    } else {
+      throw { status: 412, message: 'É necessário no mínimo um membro por familia.' };
     }
   }
+
   return db.families.create({ ...values, cityId }).then(async (family) => {
     if (values.dependents) {
       const depedentsList = values.dependents?.map((dep) => {
@@ -522,7 +526,6 @@ export const createFamilyWithDependents = async (
         family.dependents = await db.dependents.bulkCreate(depedentsList as Dependent[]);
       }
     }
-
     return family;
   });
 };
@@ -552,28 +555,38 @@ export const updateById = async (
     // The update return an array [count, item[]], so I'm destructuring to get the updated benefit
     const [, [item]] = await db.families.update(values, { where: { id }, returning: true });
 
-    if (values.dependents) {
-      const depIds = values.dependents.map((dep: Dependent) => {
-        return dep.id;
-      });
-      const familyDependents = await db.dependents.findAll({ where: { familyId: values.id as number } });
-      const dependentsToAdd = values.dependents
-        .filter((f) => !f.id)
-        .map((dep) => {
-          return { ...dep, familyId: values.id };
+    if (type === 'product') {
+      if (values.dependents) {
+        const verifyResponsible = values.dependents?.filter((f) => f.isResponsible);
+        if (verifyResponsible.length > 1)
+          throw { status: 412, message: 'Somente um membro responsável por familia é permitido.' };
+        if (verifyResponsible.length === 0)
+          throw { status: 412, message: 'É necessário um membro responsável por familia.' };
+
+        const depIds = values.dependents.map((dep: Dependent) => {
+          return dep.id;
         });
-      const dependentsToUpdate = values.dependents.filter((f) => f.id);
-      const dependentsToRemove = familyDependents.filter((f) => !depIds.includes(f.id));
+        const familyDependents = await db.dependents.findAll({ where: { familyId: values.id as number } });
+        const dependentsToAdd = values.dependents
+          .filter((f) => !f.id)
+          .map((dep) => {
+            return { ...dep, familyId: values.id };
+          });
+        const dependentsToUpdate = values.dependents.filter((f) => f.id);
+        const dependentsToRemove = familyDependents.filter((f) => !depIds.includes(f.id));
 
-      await db.dependents.bulkCreate(dependentsToAdd);
+        await db.dependents.bulkCreate(dependentsToAdd);
 
-      dependentsToUpdate.map(async (up) => {
-        if (up.id) await db.dependents.update({ ...up }, { where: { id: up.id } });
-      });
+        dependentsToUpdate.map(async (up) => {
+          if (up.id) await db.dependents.update({ ...up }, { where: { id: up.id } });
+        });
 
-      dependentsToRemove.map(async (dt) => {
-        if (dt.id) await db.dependents.destroy({ where: { id: dt.id } });
-      });
+        dependentsToRemove.map(async (dt) => {
+          if (dt.id) await db.dependents.destroy({ where: { id: dt.id } });
+        });
+      } else {
+        throw { status: 412, message: 'É necessário no mínimo um membro por familia.' };
+      }
     }
 
     return await db.families.findOne({
